% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\docType{methods}
\name{c,XPSSample-method}
\alias{c,XPSSample-method}
\title{Combine}
\usage{
\S4method{c}{XPSSample}(x, ...)
}
\arguments{
\item{x}{object of class XPSSample to be combined}

\item{...}{additional parameters}
}
\value{
XPSSample
}
\description{
Combine two XPSSample
}
