% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilities.r
\name{findXIndex}
\alias{findXIndex}
\title{findXIndex function utility}
\usage{
findXIndex(X, Value)
}
\arguments{
\item{X}{numeric vector}

\item{Value}{numeric value}
}
\value{
Returns X index
}
\description{
finds the index of X vector corresponding to a given abscissa value
}
\examples{

\dontrun{
findXIndex(X, Value)
}

}
