% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass (copy).r, R/XPSClass.r
\docType{methods}
\name{hasBoundaries}
\alias{hasBoundaries}
\title{Evaluate slot in XPSCoreLine}
\usage{
hasBoundaries(object)

hasBoundaries(object)
}
\arguments{
\item{object}{XPSCoreLine}
}
\value{
TRUE if the corresponding slot is modified.

TRUE if the corresponding slot is modified.
}
\description{
Get any parameter from slot Info of the PTRms

Get any parameter from slot Info of the PTRms
}
\details{
Any comment here.

Any comment here.
}
