% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{initialize,XPSSample-method}
\alias{initialize,XPSSample-method}
\title{'initialize'}
\usage{
\S4method{initialize}{XPSSample}(.Object, data, Project, Sample, Comments, User, Filename)
}
\arguments{
\item{.Object}{an object of class 'XPSSample'}

\item{data}{a list containing x, y original data}

\item{Project}{a character string containing information about the experiment}

\item{Sample}{a character string containing information about the analyzed sample}

\item{Comments}{a character string containing information about the experiment}

\item{User}{a character string containing information about the operator}

\item{Filename}{name of the data_file}
}
\description{
define a method 'initialize' for objects of class 'XPSSample'
}
\examples{
\dontrun{
SiOx <- initialize(.Object)
}
}
