% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{plot}
\alias{plot}
\title{'plot'}
\usage{
plot(x, y, ...)
}
\arguments{
\item{x}{XPSSample}

\item{y}{(not used) Xdata, Ydata are contained in the XPSSample object}

\item{...}{further parameters to the plot function}
}
\description{
function plot for objects of class 'XPSSample'
}
