% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{show,XPSCoreLine-method}
\alias{show,XPSCoreLine-method}
\title{show}
\usage{
\S4method{show}{XPSCoreLine}(object)
}
\arguments{
\item{object}{XPSCoreLine}
}
\description{
Method to show XPSCoreLine elements
}
\examples{
\dontrun{
 show(test.RData)
}
}
