% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{show,XPSSample-method}
\alias{show,XPSSample-method}
\title{show}
\usage{
\S4method{show}{XPSSample}(object)
}
\arguments{
\item{object}{XPSSample}
}
\description{
'show' method to print detais of objects of class XPSSample
}
\examples{
\dontrun{
 show(XPSSample.RData)
}
}
