% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass (copy).r, R/XPSClass.r
\docType{methods}
\name{sortComponents}
\alias{sortComponents}
\title{Sort Components}
\usage{
sortComponents(object)

sortComponents(object)
}
\arguments{
\item{object}{XPSCoreLine}
}
\description{
Sort Components Method for XPSCoreLine

Sort Components Method for XPSCoreLine
}
\details{
Any comment here.

Any comment here.
}
