% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSCompareGUI.r
\name{XPSCompare}
\alias{XPSCompare}
\title{Performs overlay of XPS-Spectra}
\usage{
XPSCompare()
}
\description{
Provides a userfriendly interface to select XPS-Corelines to overlay
and a selection of plotting options for a personalized data representation
This function is based on the (/code{Lattice}) Package.  No parameters passed
No parameters passed to this function
}
\examples{

\dontrun{
XPSOverlay()
}

}
