% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSDPGUI.r
\name{XPSDP}
\alias{XPSDP}
\title{Estimate the thickness of a layer by analyzing the trend of core line intensities
as a function of the tilt angle.
The Signal decay with depth as described by the Drude equation....
See Briggs...}
\usage{
XPSDP()
}
\description{
Estimate the thickness of a layer by analyzing the trend of core line intensities
as a function of the tilt angle.
The Signal decay with depth as described by the Drude equation....
See Briggs...
}
\examples{

\dontrun{
DepthPro()
}

}
