% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSExtractGUI.r
\name{XPSExtract}
\alias{XPSExtract}
\title{Extract a portion of spectrum from a XPS survey}
\usage{
XPSExtract()
}
\value{
returns the \code{object} with a coreline added.
}
\description{
Extract a portion (abscissa) from a XPS survey in a XPSSample. It use a GUI
to interact with the x-axis and to get the \code{Symbol} of the new
XPSCoreLine. XPSExtract() does not need any parameter.
}
\examples{

\dontrun{
XPSextractGUI()
}

}
