% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSFitAlgorithms.r
\name{XPSfitAlgorithms}
\alias{XPSfitAlgorithms}
\title{XPSfitAlgorithms function contains all the definition of the fit functions,
the definition of fit parameters, and the initialization of the parameter slots.}
\usage{
XPSfitAlgorithms()
}
\value{
Returns the selected fit algorithm.
}
\description{
XPSfitAlgorithms function contains all the definition of the fit functions,
the definition of fit parameters, and the initialization of the parameter slots.
}
\examples{

\dontrun{
XPSfitAlgorithms()
}

}
