% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSCoreLineInfoGUI.r, R/XPSFitParamInfoGUI.r
\docType{methods}
\name{XPSFitParamInfo}
\alias{XPSFitParamInfo}
\title{XPSFitParamInfo to describe a Core-Line best fit}
\usage{
XPSFitParamInfo()

XPSFitParamInfo()
}
\description{
XPSFitParamInfo provides the list of components and

Provides the list of components and the relative fitting parameters
for a selected XPSCoreLine. No parameters are passed to this function
}
\examples{
\dontrun{
	XPSFitParamInfo()
}

\dontrun{
XPSFitParamInfo()
}
}
