% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSMaxEnt.r
\name{XPSMaxEnt}
\alias{XPSMaxEnt}
\title{XPSMaxEnt 5-5-2020}
\usage{
XPSMaxEnt()
}
\arguments{
\item{Variance}{variance describing the degree of noise affecting the element spectra
as a function of the tilt ...}

\item{Intensity}{integral intensity of the selected corelines for each tilt angle}

\item{Conc}{element concentration value for each tilt angle}

\item{Att.L}{vector containing the element attenuation lengths.  length(L)==Nelements}

\item{Tilt}{vector containing the tilt angles  length(Tilt)==N. XPSSamples acquired for each angle}

\item{Nlayers}{number of layers (discretization) on which calculate the depth profile}

\item{LThickness}{thickness of the single layer with Nlayers*LThickness == depth modeled by MaxEnt

\dontrun{
XPSMaxEnt(Variance, Intensity, Conc, Att.L, Tilt, Nlayers, LThickness)
}}
}
\description{
Applies Max Entropy method to data extracted from XPS spectra
XPSMaxEnt may be applied for example to interpret data from ARXPS
and reconstruct a depth profile of elements in a stack of layers
XPSMaxEnt written following iterative procedure described in
A.K. Livesey, G.C. Smith, "The determination of depth profiles from angledependent XPS   (art.1)
  using maximum entropy data analysis", J. Electr. Spectr. Rel. Phen.(1994), 67, 439-461
J. Skilling, R.K. Bryan: "Max Entropy image reconstruction: general algorithm",           (art.2)
  Mon.Not. R. Astr. Soc.(1984) 211, 111
S.F. Gull and J. Skilling "Quantifed Maximum Entropy - MemSys5 Users' Manual"         (art.3)
  Maximum Entropy Data Consultants Ltd. (1999)
}
\examples{

}
