% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copy (2) of XPSMoveBaseLineGUI.r, R/Copy of
%   XPSMoveBaseLineGUI.r, R/XPSMoveBaseLineGUI.r
\name{XPSMoveBaseLine}
\alias{XPSMoveBaseLine}
\title{Modifies the BaseLine level and limits of a given Coreline}
\usage{
XPSMoveBaseLine()

XPSMoveBaseLine()

XPSMoveBaseLine()
}
\description{
Function to modify the BaseLine ends and level for a given CoreLine.
No parameters are passed to this function.

Function to modify the BaseLine ends and level for a given CoreLine.
No parameters are passed to this function.

Function to modify the BaseLine ends and level for a given CoreLine.
No parameters are passed to this function.
}
\examples{

\dontrun{
XPSMoveBaseLine()
}


\dontrun{
XPSMoveBaseLine()
}


\dontrun{
XPSMoveBaseLine()
}

}
