% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\docType{methods}
\name{XPSpkgCtrl}
\alias{XPSpkgCtrl}
\title{Function to control the attributes of an XPSSample.}
\usage{
XPSpkgCtrl(object)
}
\arguments{
\item{object}{XPSCoreLine object}
}
\description{
Function to control the attributes of an XPSSample.
}
