% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\docType{methods}
\name{XPSresidualPlot}
\alias{XPSresidualPlot}
\title{Function to plot data and the fit together with the residual difference.}
\usage{
XPSresidualPlot(object)
}
\arguments{
\item{object}{XPSCoreLine object}
}
\description{
Function to plot data and the fit together with the residual difference.
}
