% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass (copy).r, R/XPSClass.r
\docType{methods}
\name{XPSsetRSF}
\alias{XPSsetRSF}
\title{Set the RSF value for XPSCoreLine}
\usage{
XPSsetRSF(object, rsf = NULL)

XPSsetRSF(object, rsf = NULL)
}
\arguments{
\item{object}{XPSCoreLine}

\item{rsf}{RSF value}
}
\description{
Set the RSF value for XPSCoreLine

Set the RSF value for XPSCoreLine
}
\details{
Set the RSF value to the XPSCoreLine for quantification.

Set the RSF value to the XPSCoreLine for quantification.
}
