% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSprucingGUI.r
\name{XPSSprucingGUI}
\alias{XPSSprucingGUI}
\title{XPSSprucingGUI}
\usage{
XPSSprucingGUI()
}
\value{
Returns the \code{Object} with the corrected spectrum.
}
\description{
XPSSprucingGUI function to correct original XPS spectral data
}
\examples{
\dontrun{
XPSSprucingGUI()
}
}
