% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSurElmtIdentifyGUI.r
\name{XPSSurveyElementIdentify}
\alias{XPSSurveyElementIdentify}
\title{Identify Elements on the Survey}
\usage{
XPSSurveyElementIdentify()
}
\description{
GUI to Identify Elements on the Survey.
}
\details{
No parameters are passed to this function
}
\examples{

\dontrun{
XPSSurveyElementIdentify()
}

}
