% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass (copy).r, R/XPSClass.r
\docType{methods}
\name{asList}
\alias{asList}
\title{asList}
\usage{
asList(from, select = "all")

asList(from, select = "all")
}
\arguments{
\item{from}{XPSCoreLine}

\item{select}{one or few or \code{"all"} (default) of \code{c("MAIN", "RTF", "BASE", "COMPONENTS", "FIT")}}
}
\value{
A list

A list
}
\description{
asList Method for XPSCoreLine

asList Method for XPSCoreLine
}
\details{
Transform an XPSCoreLine as list. Mainly used for plot function.

Transform an XPSCoreLine as list. Mainly used for plot function.
}
