% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass (copy).r, R/XPSClass.r
\docType{methods}
\name{c,XPSSample-method}
\alias{c,XPSSample-method}
\title{Combine}
\usage{
\S4method{c}{XPSSample}(x, ...)

\S4method{c}{XPSSample}(x, ...)
}
\value{
XPSSample

XPSSample
}
\description{
Combine two XPSSample

Combine two XPSSample
}
