% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass.r
\name{plot}
\alias{plot}
\title{Plot method for XPSCoreLine or XPSSample objects}
\usage{
plot(x, y, ...)

plot(x, y, ...)
}
\arguments{
\item{x}{XPSSample}

\item{y}{missing parameter

\code{TRUE}}

\item{...}{further parameters to the plot function}
}
\description{
Plot XPSCoreLine or XPSSample objects.
}
\details{
The normal way to plot \code{XPS} objects is to use the \code{plot} method.
In the case of \code{XPSCoreLine} object it is a wrapper for
\code{\link{matplot}}. The data are transformed to \code{matrix} for \code{\link{matplot}}
unprocessed CoreLine. For processed one it plots the RegionToFit defined by
the Boundaries extremes, the baseline, any FitComponent and Fit if they are present.
}
\seealso{
\code{\link{matplot}}, \code{\link{par}}
S4method plot
}
\author{
R. Canteri
}
