% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass (copy).r, R/XPSClass.r
\docType{methods}
\name{show,XPSCoreLine-method}
\alias{show,XPSCoreLine-method}
\title{Show}
\usage{
\S4method{show}{XPSCoreLine}(object)

\S4method{show}{XPSCoreLine}(object)
}
\arguments{
\item{object}{XPSCoreLine}
}
\description{
Show Method for XPSCoreLine

Show Method for XPSCoreLine
}
\details{
Any comment here.

Any comment here.
}
