% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSClass (copy).r, R/XPSClass.r
\docType{methods}
\name{show,XPSSample-method}
\alias{show,XPSSample-method}
\title{Show}
\usage{
\S4method{show}{XPSSample}(object)

\S4method{show}{XPSSample}(object)
}
\arguments{
\item{object}{XPSSample}
}
\description{
Show Method for XPSSample

Show Method for XPSSample
}
\details{
Any comment here.

Any comment here.
}
