% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSAnalysisGUI.r
\name{XPSAnalysis}
\alias{XPSAnalysis}
\title{GUI to analyze XPS-Corelines}
\usage{
XPSAnalysis()
}
\value{
Returns the processed \code{object}.
}
\description{
Interactive GUI to add BaseLines and Fitting components to a CoreLine
No parameter passed to this function
}
\examples{

\dontrun{
XPSAnalysis()
}

}
