% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSImportAsciiGUI.r
\name{XPSImport.Ascii}
\alias{XPSImport.Ascii}
\title{XPSImport.Ascii function}
\usage{
XPSImport.Ascii()
}
\value{
Return the processed \code{object}.

\dontrun{
XPSImport.Ascii()
}
}
\description{
Function allowing import of textual (ascii) file data
Options are available to account for header and data separator
Options are also provided to store data in a XPS-Sample data-structure.
No parameters are passed to this function
}
