% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSSprucingGUI.r
\name{XPSSprucingGUI}
\alias{XPSSprucingGUI}
\title{Sprucing XPSSample data}
\usage{
XPSSprucingGUI()
}
\value{
Returns the \code{Object} with corrected Survey spectrum.
}
\description{
GUI to correct original XPS spectral data
}
\examples{

\dontrun{
XPSSprucingGUI()
}

}
